+++ OpenMEEG/include/MeshIOs/Unavailable.h
@@ -46,7 +46,7 @@ knowledge of the CeCILL-B license and that you accept its terms.
 #include <string>
 
 #include <om_utils.h>
-#include <mesh_ios.h>
+#include <MeshIO.h>
 
 #ifdef USE_GIFTI
 extern "C" {
