+++ OpenMEEG/include/MeshIOs/off.h
@@ -46,7 +46,7 @@ knowledge of the CeCILL-B license and that you accept its terms.
 #include <string>
 
 #include <om_utils.h>
-#include <mesh_ios.h>
+#include <MeshIO.h>
 #include <GeometryExceptions.H>
 
 namespace OpenMEEG::MeshIOs {
