+++ OpenMEEG/include/MeshIOs/tri.h
@@ -43,7 +43,7 @@ knowledge of the CeCILL-B license and that you accept its terms.
 #include <string>
 
 #include <om_utils.h>
-#include <mesh_ios.h>
+#include <MeshIO.h>
 
 namespace OpenMEEG::MeshIOs {
